refactor(hooks): remove legacy useCDSHooks.js shim
`frontend/src/hooks/useCDSHooks.js` was a one-line re-export shim
(`export * from './cds/useCDSHooks'`) kept for backward compat after
the canonical location moved to `hooks/cds/useCDSHooks.js`. Five
consumers still imported through the shim path, which means every
contributor working in this area has to discover the duplicate file
exists, follow the re-export, and confirm both files have the same
exports before they trust anything they're reading.

Update the five remaining import sites to point at the canonical
`hooks/cds/useCDSHooks` path. Delete the shim. No behavior change —
the underlying hook implementation is unchanged.

Verified with `grep -rn "from.*hooks/useCDSHooks'"` that no other
consumer references the legacy path. Test surface unaffected: the new
suites I added in PR #117 already imported from the canonical
location, and the existing-orders-list path that was always at the
shim was the only behavior-touching call site (now spec-correct via
the same hook).
