--- frontend/src/hooks/useCDSHooks.js
@@ -1,2 +1,0 @@
-// Re-export from new location for backwards compatibility
-export * from './cds/useCDSHooks';
